feat(pkg): init forecast pkg
